requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
environ==1.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
PyJWT==2.9.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
